add data-processing modules for `phf`
- brace-initialization featured numpy-like; in order to conveniently use brace init (for function-call of the newly set modules) without stating the defined type (i.e., `hfarray_cxx<T>` in this case), it must NOT be nvcc to be the primary/host compiler, which mandates a compiler firewall & an adequate C++ compiler (C++17 support is otherwise put for this project).
- misc: 1) hf.hh -> hfclass.hh; 2) replace using modules; 3) remove used 8-byte intermediate in phf

+++ include/hf/hfcodec.hh
@@ -28,8 +28,8 @@ void hf_encode_coarse_rev2(
 
 template <typename T, typename H, typename M>
 void hf_decode_coarse(
-    H* d_bitstream, uint8_t* d_revbook, int const revbook_nbyte, M* d_par_nbit,
-    M* d_par_entry, int const sublen, int const pardeg, T* out_decompressed,
+    H* d_bitstream, uint8_t* d_revbook, size_t const revbook_nbyte, M* d_par_nbit,
+    M* d_par_entry, size_t const sublen, size_t const pardeg, T* out_decompressed,
     float* time_lossless, void* stream);
 
 }  // namespace psz
